+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -38,6 +38,7 @@ import com.onlyoffice.service.documenteditor.config.ConfigService;
 import onlyoffice.managers.auth.AuthContext;
 import com.atlassian.confluence.pages.Attachment;
 import onlyoffice.sdk.manager.document.DocumentManager;
+import onlyoffice.sdk.manager.security.JwtManager;
 import onlyoffice.sdk.manager.url.UrlManager;
 import onlyoffice.utils.attachment.AttachmentUtil;
 import org.apache.log4j.LogManager;
@@ -68,13 +69,14 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
     private final AttachmentUtil attachmentUtil;
     private final ConfigService configService;
     private final SettingsManager settingsManager;
+    private final JwtManager jwtManager;
 
     private final LocaleManager localeManager;
 
     public OnlyOfficeEditorServlet(final I18nResolver i18n, final UrlManager urlManager, final AuthContext authContext,
                                    final DocumentManager documentManager, final AttachmentUtil attachmentUtil,
                                    final ConfigService configService, final SettingsManager settingsManager,
-                                   final LocaleManager localeManager) {
+                                   final JwtManager jwtManager, final LocaleManager localeManager) {
         this.i18n = i18n;
         this.urlManager = urlManager;
         this.authContext = authContext;
@@ -82,6 +84,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         this.attachmentUtil = attachmentUtil;
         this.configService = configService;
         this.settingsManager = settingsManager;
+        this.jwtManager = jwtManager;
         this.localeManager = localeManager;
     }
 
@@ -111,7 +114,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
                 String extension = fileExt == null
                         || !fileExt.equals("xlsx")
                         && !fileExt.equals("pptx")
-                        && !fileExt.equals("docxf")
+                        && !fileExt.equals("pdf")
                         ? "docx" : fileExt.trim();
 
                 String name = fileName == null || fileName.equals("")
@@ -182,8 +185,19 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
                         request.getHeader("USER-AGENT")
                 );
 
+                if (modeString != null
+                        && modeString.equals("fillForms")
+                        && config.getDocument().getPermissions().getFillForms()
+                ) {
+                    config.getDocument().getPermissions().setEdit(false);
+                }
+
                 config.getEditorConfig().setLang(localeManager.getLocale(user).toLanguageTag());
                 config.getEditorConfig().setActionLink(actionData);
+
+                if (settingsManager.isSecurityEnabled()) {
+                    config.setToken(jwtManager.createToken(config));
+                }
 
                 ObjectMapper mapper = createObjectMapper();